define global state
